crawler/requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
idna==2.10
PyMySQL==1.0.2
PySocks==1.7.1
requests==2.25.1
soupsieve==2.2.1
termcolor==1.1.0
urllib3==1.26.5
